worksheets are now user specific, created page to see all worksheets

+++ main/views.py
@@ -11,8 +11,10 @@ def home(response):
         form = CreateNewWorksheet(response.POST)
 
         if form.is_valid():
-            worksheet = Worksheet(name=form.cleaned_data["name"])
+            n = form.cleaned_data["name"]
+            worksheet = Worksheet(name=n)
             worksheet.save()
+            response.user.worksheet.add(worksheet)
 
         return HttpResponseRedirect("tracker/%i" % worksheet.id)
     else:
@@ -28,32 +30,42 @@ def dashboard(response):
 
 
 def user_tracker(response, id):
-    ws = Worksheet.objects.get(id=id)
-    last_entry = len(ws.entry_set.all())
-    if response.method == "POST":
-        if response.POST.get("start_new"):
-            if (last_entry == 0):
-                entry = ws.entry_set.create()
-                entry.store = response.POST.get("btnradio")
-                entry.operation = response.POST.get("operation_radio")
-                entry.save()
-            else:
-                entry = ws.entry_set.order_by('start_time')[last_entry - 1]
-                entry.end_time = timezone.now()
-                entry.save()
-                new_entry = ws.entry_set.create()
-                new_entry.operation = response.POST.get("operation_radio")
-                new_entry.store = response.POST.get("btnradio")
-                new_entry.save()
+    try:
+        ws = Worksheet.objects.get(id=id)
+    except:
+        form = CreateNewWorksheet()
+        return render(response, 'home.html', {"form": form})
+        
+    if ws in response.user.worksheet.all():
+        
+        last_entry = len(ws.entry_set.all())
+        if response.method == "POST":
+            if response.POST.get("start_new"):
+                if (last_entry == 0):
+                    entry = ws.entry_set.create()
+                    entry.store = response.POST.get("btnradio")
+                    entry.operation = response.POST.get("operation_radio")
+                    entry.save()
+                else:
+                    entry = ws.entry_set.order_by('start_time')[last_entry - 1]
+                    entry.end_time = timezone.now()
+                    entry.save()
+                    new_entry = ws.entry_set.create()
+                    new_entry.operation = response.POST.get("operation_radio")
+                    new_entry.store = response.POST.get("btnradio")
+                    new_entry.save()
 
-        elif response.POST.get("delete"):
-            #if delete on no entries, delete worksheet and return to home
-            if (last_entry == 0):
-                # ws.delete()
-                # return HttpResponseRedirect("/")
-                pass
-            else:
-                entry = ws.entry_set.order_by('start_time')[last_entry - 1]
-                entry.delete()
+            elif response.POST.get("delete"):
+                #if delete on no entries, delete worksheet and return to home
+                if (last_entry == 0):
+                    # ws.delete()
+                    # return HttpResponseRedirect("/")
+                    pass
+                else:
+                    entry = ws.entry_set.order_by('start_time')[last_entry - 1]
+                    entry.delete()
+        return render(response, 'tracker.html', {"ws": ws})
+    return render(response, "worksheets.html", {"ws": ws})
 
-    return render(response, 'tracker.html', {"ws": ws})
+def worksheets(response):
+    return render(response, "worksheets.html", {})
